Rename data store wrapper to ananke data store
