--- app/src/main/java/com/carkzis/ananke/data/DataStoreWrapper.kt
@@ -1,8 +1,0 @@
-package com.carkzis.ananke.data
-
-import kotlinx.coroutines.flow.Flow
-
-interface DataStoreWrapper {
-    val data: Flow<String?>
-    suspend fun setCurrentGameId(gameId: String)
-}
