+++ app/src/test/java/com/carkzis/ananke/testdoubles/FailingAnankeDataStore.kt
@@ -1,11 +1,11 @@
 package com.carkzis.ananke.testdoubles
 
-import com.carkzis.ananke.data.DataStoreWrapper
+import com.carkzis.ananke.data.AnankeDataStore
 import com.carkzis.ananke.ui.screens.nugame.EnterGameFailedException
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 
-class FailingDataStore : DataStoreWrapper {
+class FailingAnankeDataStore : AnankeDataStore {
     override val data: Flow<String?> = flow {}
     override suspend fun setCurrentGameId(gameId: String) {
         throw EnterGameFailedException()
